requirements.txt
# Machine Learning
tensorflow==2.20.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
pillow==10.0.0
opencv-python==4.8.0.74
mlflow==2.7.1

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality
flake8==6.1.0
black==23.7.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1